Move callback_executor into postgresql package
