+++ patroni/postgresql/callback_executor.py
